Fixed #29
I have added an extra "filter" when checking if the href found in the token starts with "/wiki/"; the program now also checks if the found href contains either "#" or ":", which either stand for a particular part of a possibly already visited article, or for categories and special pages, such as user profiles. These sorts of pages are of no interest to us when trying to determine the main structure of Wikipedia.